+++ CHANGELOG.md
@@ -179,3 +179,4 @@
 - 2023-11-21: Features contacts, chat, Makefile, campaign, +4
 - 2023-11-21: Wire lead, qualify lead, qualify, mod, +6
 - 2023-11-22: Builder icp, discover, deal, ready pool, +3
+- 2023-11-23: Shutdown ready pool, qualify lead, lead, users, +5